[MOV] tests: move clickable cells tests to their own file
Task: 4630549
